scripts/requirements.txt
pyserial~=3.0.1
